% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{fake_fish_creator}
\alias{fake_fish_creator}
\title{A function to create fake Length/Age relationship for fish}
\usage{
fake_fish_creator(n, maxA, noiseK, noiseL, K, Linf)
}
\arguments{
\item{n}{Number of fish captured}

\item{maxA}{Age of the oldest fish captured (years)}

\item{noiseK}{Standard deviation of the population growth rate (mm/years)}

\item{noiseL}{Standard deviation of the population maximum size (mm)}

\item{K}{Mean of the population growth rate (mm/years)}

\item{Linf}{Mean of the population maximum size (mm)}
}
\value{
A dataframe containing the simulated Age(year) and Length(mm) along with a plot presenting Length(mm) as a function of age(year).
}
\description{
This function allows you create fake data, with added noise.
}
\examples{
function(12,32,0,01,10,0.5,1200)
}
\keyword{Bertalanffy,}
\keyword{Fish,}
\keyword{Von}
\keyword{simulation}
