% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKest.R
\name{LKest}
\alias{LKest}
\title{Linf and K estimator for a Von Bertalanffy Age/Length relationship}
\usage{
LKest(Length, Age, plotVBGF, t0_estim, na.rm, boot, boundary)
}
\arguments{
\item{Length}{a vector comtaining the length measurements}

\item{Age}{a vector comtaining the age measurements}

\item{plotVBGF}{if \code{TRUE} plot of the VB model, \code{FALSE} does not provide a plot}

\item{t0_estim}{the value of t0 for the model (0 if no better guess is usual)}

\item{na.rm}{if \code{TRUE} remove NA in the first and second arguments, \code{FALSE} else}

\item{boot}{if \code{TRUE} plot 1000 parameters estimations value issued through a boostrap procedure (the corresponding 95 percent confidence interval is plotted as a dotted line)}

\item{boundary}{if TRUE, value of K bounded between 0 and 1}
}
\value{
A list containing (i) the Von Bertalanffy growth curve fitted values, (ii) the Von Bertalanffy growth curve parameters , (iii) the confidence interval fitted values obtained using the appraoch described in Kimura (1980), (iv) the estimated Von Bertalanffy parameters using the approach described in Kimura (1980), and (v) the r-squared value.
}
\description{
This function allows you to estimates the growth rate \code{K} and the theoretical maximum size \code{Linf} from a dataset made of \code{Age} and \code{length} measurements using teh method described in Kimura D.K.,(1980) Likelihood methods for the von Bertalanffy growth curve. Fishery Bulletin 77: 765–776.
}
\examples{
no example so far, chill
}
\keyword{Bertalanffy,}
\keyword{Fish,}
\keyword{K,}
\keyword{Kimura,}
\keyword{Linf,}
\keyword{Von}
\keyword{boostrap}
\keyword{estination,}
